requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# HTTP clients
requests==2.31.0
aiohttp==3.9.1

# Data science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.3.2

# NASA data
netCDF4==1.6.4

# Async support
asyncio-mqtt==0.16.1

# CORS
python-multipart==0.0.6